Duration Cookies not working due to overloading issue

+++ framework/src/play/src/main/scala/play/core/j/JavaHelpers.scala
@@ -33,7 +33,7 @@ trait JavaHelpers {
     case result: PlainResult => {
       val wResult = result.withHeaders(javaContext.response.getHeaders.asScala.toSeq: _*)
         .withCookies((javaContext.response.cookies.asScala.toSeq map { c => Cookie(c.name, c.value,
-          if (c.maxAge == null) None else Some(c.maxAge), c.path, Option(c.domain), c.secure, c.httpOnly) }): _*)
+          if (c.maxAge == null) None else Some(c.maxAge.toInt), c.path, Option(c.domain), c.secure, c.httpOnly) }): _*)
 
       if (javaContext.session.isDirty && javaContext.flash.isDirty) {
         wResult.withSession(Session(javaContext.session.asScala.toMap)).flashing(Flash(javaContext.flash.asScala.toMap))
